Refactored SectionsContent code by adding a Section & Question components
